Update calc.js
Version 2.1 of the calculator javascript code, it mainly entails a bug patch that fixes an issue that resulted in incorrect results if the user attempted to do consecutive arithmetic operations without pressing enter or equals (for example: 1 + 1 + 1 would equal 2 instead of 3, however if you did 1 + 1 = 2 + 1 = you would get the correct answer of 3). This bug patch will now allow the user to do these consecutive arithmetic operations as described above and get the correct results. 

Also in this code update you may notice some other incomplete code (that has been commented out for the time being) for a future update where I plan to format the numbers with commas as needed in the display (for example 1,000 or 10,000). This functionality is not completed yet so all code relating to it as has been commented out but left in,